requirements.txt
aiogram>3
python-socketio[client]
colorama